Remove circular dependencies for a happy, healthy BSI
