+++ consts.js
@@ -1,0 +1,17 @@
+export const COURSE_OFFERING = 3;
+
+export const RECORD = {
+	ORG_UNIT_ID: 0,
+	USER_ID: 1,
+	ROLE_ID: 2,
+	OVERDUE: 3,
+	CURRENT_FINAL_GRADE: 4,
+	TIME_IN_CONTENT: 5,
+	COURSE_LAST_ACCESS: 6
+};
+export const USER = {
+	ID: 0,
+	FIRST_NAME: 1,
+	LAST_NAME: 2,
+	USERNAME: 3
+};
