+++ model/data.js
@@ -1,28 +1,10 @@
 import { action, autorun, computed, decorate, observable } from 'mobx';
+import { COURSE_OFFERING, RECORD, USER } from '../consts';
 import { OrgUnitSelectorFilter, RoleSelectorFilter, SemesterSelectorFilter } from './selectorFilters.js';
 import { CardFilter } from './cardFilter.js';
 import { fetchCachedChildren } from './lms.js';
 import { TABLE_USER } from '../components/users-table';
 import { Tree } from '../components/tree-filter';
-
-export const COURSE_OFFERING = 3;
-
-export const RECORD = {
-	ORG_UNIT_ID: 0,
-	USER_ID: 1,
-	ROLE_ID: 2,
-	OVERDUE: 3,
-	CURRENT_FINAL_GRADE: 4,
-	TIME_IN_CONTENT: 5,
-	COURSE_LAST_ACCESS: 6
-};
-
-export const USER = {
-	ID: 0,
-	FIRST_NAME: 1,
-	LAST_NAME: 2,
-	USERNAME: 3
-};
 
 function unique(array) {
 	return [...new Set(array)];
